examples/requirements.txt
jinja2
jinja2_markdown
pyyaml
